+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.5.0]
+- Make possible to style spec. tile icons
+- Allows incomplete configuration if the tile is disabled
+- Show language flag
+
 ## [1.4.3]
 - Display placeholder types in notification form